third_party/blink/web_tests/platform/generic/fast/filesystem/simple-required-arguments-getfile-expected.txt

DirectoryEntry required arguments test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

webkitRequestFileSystem(TEMPORARY, 100, successCallback, errorCallback);
Successfully obtained TEMPORARY FileSystem:file__0:Temporary
root = fileSystem.root
PASS root.getFile() threw exception TypeError: Failed to execute 'getFile' on 'DirectoryEntry': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
